+++ functions/util/firebaseConfig.js
@@ -1,0 +1,10 @@
+module.exports = {
+    apiKey: "AIzaSyAUgEYIfl619L0JzRkzZQny350mYmcEO6Q",
+    authDomain: "memories-284f9.firebaseapp.com",
+    databaseURL: "https://memories-284f9-default-rtdb.firebaseio.com",
+    projectId: "memories-284f9",
+    storageBucket: "memories-284f9.appspot.com",
+    messagingSenderId: "692138113576",
+    appId: "1:692138113576:web:68d235f7b3970835a7531d",
+    measurementId: "G-N2BB27WH1Z"
+  };
